Completed my solution to flex exercise 5
I am committing this before looking at the solution. My answer consists of a flex box containing the icon
and another flex box that contains the rest of the elements. The second flexbox has the header at flex 1
so it moves the x button to the far right, and a max width is put on the text to match solution image